Feat: Add Ergonomical and Self Documenting API
The idea is for the API being able to generate itself according to
Actions.
Cases: If a new endpoint is added it will automatically be recognized
and reported with a simple GET on the root endpoint.

When an endpoint is doing something, this should self document what each
endpoint does both to who writes the code and to whoever access the API